+++ example/lib/src/common/widget/app.dart
@@ -18,16 +18,23 @@ class _AppState extends State<App> with RouteStateMixin {
   final GlobalKey<State<StatefulWidget>> _preserveKey =
       GlobalKey<State<StatefulWidget>>();
 
+  late final RouterConfig<Object> _routerConfig;
+
   @override
-  Widget build(BuildContext context) => MaterialApp(
+  void initState() {
+    super.initState();
+    _routerConfig = Elixir.router(
+      routes: routes,
+      initialStack: initialPages,
+      guards: guards,
+    );
+  }
+
+  @override
+  Widget build(BuildContext context) => MaterialApp.router(
     key: _preserveKey,
     title: 'Declarative Navigation',
     debugShowCheckedModeBanner: false,
-    builder:
-        (context, _) => Elixir.controlled(
-          controller: ValueNotifier(initialPages),
-          guards: guards,
-          routes: routes,
-        ),
+    routerConfig: _routerConfig,
   );
 }