+++ lib/elixir.dart
@@ -1,5 +1,5 @@
 export 'src/navigator/elixir_page.dart';
-export 'src/navigator/navigator.dart';
 export 'src/navigator/observer.dart';
 export 'src/navigator/types.dart';
+export 'src/router/router.dart';
 export 'src/util/context_extension.dart';